Update Makefiles for CBMC proofs
