--- .cbmc-batch/jobs/Makefile
@@ -1,2 +1,0 @@
-goto: proofs.goto
-	goto-instrument proofs.goto proofs.goto
