+++ .cbmc-batch/jobs/aws_array_list_init_static/cbmc-batch.yaml
@@ -1,4 +1,4 @@
 jobos: ubuntu16
-cbmcflags: "--bounds-check;--pointer-check;--function;aws_array_list_init_static_verify"
-goto: proofs.goto
+cbmcflags: "--bounds-check;--pointer-check;--function;aws_array_list_init_static_harness"
+goto: aws_array_list_init_static_harness.goto
 expected: "SUCCESSFUL"
