+++ .cbmc-batch/jobs/aws_byte_buf_append/cbmc-batch.yaml
@@ -1,4 +1,4 @@
 jobos: ubuntu16
-cbmcflags: "--bounds-check;--pointer-check;--function;aws_byte_buf_append_verify;--unwinding-assertions;--unwindset;memcpy.0:33,__builtin___memcpy_chk.0:33"
-goto: proofs.goto
+cbmcflags: "--bounds-check;--pointer-check;--function;aws_byte_buf_append_harness;--unwinding-assertions;--unwindset;memcpy.0:33,__builtin___memcpy_chk.0:33"
+goto: aws_byte_buf_append_harness.goto
 expected: "SUCCESSFUL"
