+++ .cbmc-batch/jobs/aws_byte_buf_from_c_str/cbmc-batch.yaml
@@ -1,4 +1,4 @@
 jobos: ubuntu16
-cbmcflags: "--bounds-check;--pointer-check;--function;aws_byte_buf_from_c_str_verify;--unwindset;strlen.0:33;--unwinding-assertions"
-goto: proofs.goto
+cbmcflags: "--bounds-check;--pointer-check;--function;aws_byte_buf_from_c_str_harness;--unwindset;strlen.0:33;--unwinding-assertions"
+goto: aws_byte_buf_from_c_str_harness.goto
 expected: "SUCCESSFUL"
